requirements.txt
flask
gunicorn
pillow
rembg
base64
uuid